+++ apps/main-web/src/app/(auth)/sign-up/page.tsx
@@ -1,7 +1,7 @@
 import dynamic from "next/dynamic";
 
 const SignUpForm = dynamic(
-  () => import("../../../components/layouts/signup/SignUpForm"),
+  () => import("../../../components/forms/SignUpForm"),
   {
     ssr: false,
   },
