+++ apps/main-web/src/components/common/feedcard/CardMainContent.tsx
