+++ apps/main-web/src/components/common/feedcard/CardMainImage.tsx
