+++ apps/main-web/src/components/forms/UserInfoForm.tsx
@@ -6,22 +6,21 @@ import {
 } from "@repo/ui/components/ui/form";
 import { FormProvider, useForm } from "react-hook-form";
 import {
+  checkCredentialsAvailabilityApi,
+  registerUserApi,
+} from "../../actions/auth/sign-up";
+import {
   RegisterUserInfo,
   UserInfoSchema,
   UserInfoType,
-} from "../../../types/auth/signup";
-import {
-  checkCredentialsAvailabilityApi,
-  registerUserApi,
-} from "../../../actions/auth/sign-up";
+} from "../../types/auth/signup";
 
-import { Button } from "@repo/ui/components/ui/button";
-import { CommonResponse } from "../../../types/responseType";
-import { Input } from "@repo/ui/components/ui/input";
-import ProgressBar from "./ProgressBar";
-import React from "react";
-import { useRouter } from "next/navigation";
 import { zodResolver } from "@hookform/resolvers/zod";
+import { Button } from "@repo/ui/components/ui/button";
+import { Input } from "@repo/ui/components/ui/input";
+import { useRouter } from "next/navigation";
+import { CommonResponse } from "../../types/responseType";
+import ProgressBar from "../pages/signup/ProgressBar";
 
 interface UserInfoProps {
   terms: {
@@ -44,7 +43,7 @@ const formInputs: Array<{
   { name: "nickname", label: "닉네임", type: "text" },
 ];
 
-export default function UserInfo({
+export default function UserInfoForm({
   terms,
   email,
   emailVerificationCode,
