+++ apps/main-web/src/components/icons/CatJelly.tsx
