+++ apps/main-web/src/components/icons/LookidsLogo.tsx
