+++ apps/main-web/src/components/icons/bottomNavBar/BottomAddIcon.tsx
