+++ apps/main-web/src/components/icons/bottomNavBar/BottomDiaryIcon.tsx
