+++ apps/main-web/src/components/icons/bottomNavBar/BottomHomeIcon.tsx
@@ -1,4 +1,4 @@
-import { IconActiveType } from "../../types/bar/IconActiveType";
+import { IconActiveType } from "../../../types/bar/IconActiveType";
 
 export default function BottomHomeIcon({ isActive = false }: IconActiveType) {
   return (
