+++ apps/main-web/src/components/icons/bottomNavBar/BottomMapIcon.tsx
