+++ apps/main-web/src/components/icons/bottomNavBar/BottomMyPageIcon.tsx
