+++ apps/main-web/src/components/icons/topNavBar/MainHamburger.tsx
