+++ apps/main-web/src/components/icons/topNavBar/MainTopDiary.tsx
