+++ apps/main-web/src/components/icons/topNavBar/MainTopLogo.tsx
