+++ apps/main-web/src/components/icons/topNavBar/MainTopSearch.tsx
