+++ apps/main-web/src/components/pages/main/SwiperCarouselCard.tsx
