+++ apps/main-web/src/components/pages/signup/ProgressBar.tsx
