+++ apps/main-web/src/components/pages/signup/TermList/Term1.tsx
