+++ apps/main-web/src/components/pages/signup/TermList/Term2.tsx
