+++ apps/main-web/src/components/ui/SignUpHeader.tsx
@@ -1,14 +1,13 @@
 import { ChevronLeft } from "lucide-react";
-import React from "react";
 
-export default function SignUpHeader() {
+export default function CommonHeader({ title }: { title: string }) {
   return (
     <section className="mt-[52px] h-12 flex items-center relative">
       <ChevronLeft
         className="absolute left-3"
         onClick={() => window.history.back()}
       />
-      <p className="font-semibold text-center flex-1">회원가입</p>
+      <p className="font-semibold text-center flex-1">{title}</p>
     </section>
   );
 }
