+++ apps/main-web/src/components/utils/sample.ts
@@ -1,0 +1,3 @@
+// Scroll 관련 함수 등 모든 곳에서 사용할 수 있는 함수들을 모아둔다.
+// 이 파일은 다른 파일에서 import하여 사용한다.
+// 이 파일은 다른 파일에서 import하여 사용한다.
