+++ apps/main-web/src/hooks/customhook.ts
@@ -1,0 +1,1 @@
+// 이 파일은 custom hook을 만들어서 사용하는 예제입니다.
